Refactor fitness app structure by removing ActivityCard and introducing Activity and Workout models
- Deleted the ActivityCard component and moved its functionality to the new Activity and Workout model files.
- Created HomeViewModel to manage mock data for activities and workouts.
- Added new views: HomeView for displaying user statistics and activity cards, ProgressCircleView for visual progress representation, and WorkoutCard for displaying workout details.
- Enhanced HomeView layout to incorporate new components and improve user experience.
