+++ fitness-app/Home/Models/Activity.swift
@@ -1,0 +1,17 @@
+//
+//  Activity.swift
+//  fitness-app
+//
+//  Created by Patrick Niyo on 11/01/2025.
+//
+
+import SwiftUI
+
+struct Activity {
+    let id: Int
+    let title: String
+    let subtitle: String
+    let image: String
+    let tinColor: Color
+    let amount: String
+}
