+++ fitness-app/Home/Models/Workout.swift
@@ -1,0 +1,18 @@
+//
+//  Workout.swift
+//  fitness-app
+//
+//  Created by Patrick Niyo on 11/01/2025.
+//
+
+import SwiftUI
+
+struct Workout {
+    let id: Int
+    let title: String
+    let image: String
+    let duration: String
+    let date: String
+    let calories: String
+    let tinColor: Color
+}
