+++ fitness-app/Home/Views/ProgressCircleView.swift
