+++ fitness-app/Home/Views/WorkoutCard.swift
@@ -7,15 +7,7 @@
 
 import SwiftUI
 
-struct Workout {
-    let id: Int
-    let title: String
-    let image: String
-    let duration: String
-    let date: String
-    let calories: String
-    let tinColor: Color
-}
+
 struct WorkoutCard: View {
     @State var workout: Workout
